# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ConnectionRead": {
        "description": "Safe projection of a connection. Has no field capable of holding a secret.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "database": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Database"
          },
          "db_type": {
            "$ref": "#/components/schemas/DbType"
          },
          "host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Host"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_test_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Test Error"
          },
          "last_tested_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Tested At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "paused": {
            "default": false,
            "title": "Paused",
            "type": "boolean"
          },
          "port": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Port"
          },
          "sqlite_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sqlite Path"
          },
          "ssl_mode": {
            "$ref": "#/components/schemas/SslMode"
          },
          "ssl_root_cert": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ssl Root Cert"
          },
          "status": {
            "$ref": "#/components/schemas/ConnectionStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          }
        },
        "required": [
          "id",
          "name",
          "db_type",
          "host",
          "port",
          "database",
          "username",
          "sqlite_path",
          "ssl_mode",
          "ssl_root_cert",
          "status",
          "last_tested_at",
          "last_test_error",
          "created_at",
          "updated_at"
        ],
        "title": "ConnectionRead",
        "type": "object"
      },
      "ConnectionStatus": {
        "enum": [
          "untested",
          "ok",
          "failed"
        ],
        "title": "ConnectionStatus",
        "type": "string"
      },
...
      "DbType": {
        "enum": [
          "postgres",
          "mysql",
          "sqlite"
        ],
        "title": "DbType",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "Schema-agnostic backend for saved, read-only SQL exposed as chart-ready JSON. All fraud logic lives in the SQL you save.",
    "title": "Fraud Analyzer Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/connections": {
      "get": {
        "description": "List every connection. Credentials are never included.",
        "operationId": "list_connections_connections_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConnectionRead"
                  },
                  "title": "Response List Connections Connections Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          }
        },
        "summary": "List Connections",
        "tags": [
          "connections"
        ]
      },
      "post": {
        "description": "Create a connection profile and test it immediately.\n\nA failing test still saves the profile, with ``status=\"failed\"`` and the\nerror attached, so credentials can be corrected without re-entering\neverything.",
        "operationId": "create_connection_connections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionCreateResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          }
        },
        "summary": "Create Connection",
        "tags": [
          "connections"
        ]
      }
    },
...
    "/connections/{connection_id}/query/preview": {
      "post": {
        "description": "Run ad-hoc SQL without saving it, for a 'try before you save' UX.\n\nGoes through the same guard as everything else and is capped aggressively,\nsince this is exploratory only. Nothing is *persisted* - no saved query,\nno cache entry, no flagged rows - but the run itself is logged, with the\ncaller's id and the connection it touched. Exploratory or not, this is\nanalyst-authored SQL against a customer's database, which is precisely the\nthing the execution log exists to attribute; \"we did not save the query\"\nis not a reason to leave no record that it ran. There is no saved query to\nhang the row on, so it carries ``connection_id`` instead - see\n``QueryExecutionLog.connection_id``.",
        "operationId": "preview_query_connections__connection_id__query_preview_post",
        "parameters": [
          {
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "title": "Connection Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          }
        },
        "summary": "Preview Query",
        "tags": [
          "queries"
        ]
      }
    },
...
  }
}